7219 all the things update sitemap & route remove dd refs from download page move dd files from download folders to their own rename folders & files to remove casing (just in case) update & simplify css
