+++ scripts/sitemaps/pages.js
@@ -21,7 +21,8 @@ const routes = [
     '/state',
     '/recipient',
     '/disaster/covid-19',
-    '/disaster/covid-19/data-sources'
+    '/disaster/covid-19/data-sources',
+    '/data-dictionary'
 ];
 
 const recipientRequestObject = {
