+++ src/_scss/pages/data-dictionary/_dictionary-table.scss
