+++ src/_scss/pages/data-dictionary/_search-bar.scss
