+++ src/_scss/pages/data-dictionary/data-dictionary-page.scss
@@ -1,0 +1,60 @@
+.usa-da-data-dictionary-page {
+	@import 'all';
+	@import 'layouts/default/default';
+	@import 'layouts/default/stickyHeader/header';
+
+	#main-content {
+		@import '../../mixins/fullSectionWrap';
+		@import '../../components/pageLoading';
+		@include fullSectionWrap(($global-mrg * 2), ($global-mrg * 2));
+
+		background-color: $color-white;
+		box-shadow: $container-shadow;
+		border-top: 1px solid $color-gray-border;
+		border-right: 1px solid $color-gray-border;
+		border-bottom: 1px solid $color-gray-border;
+		padding: rem(30) rem(20);
+
+		.data-dictionary {
+			@include media($medium-screen) {
+				padding: 0 rem(30);
+			}
+			.data-dictionary__title {
+				margin-top: 0;
+				font-weight: 400;
+			}
+			.data-dictionary__intro {
+				font-size: $small-font-size;
+			}
+			.data-dictionary__search-download {
+				@include display(flex);
+				@include justify-content(space-between);
+			}
+			.data-dictionary__download {
+				@include align-self(center);
+				padding-bottom: rem(5);
+				.data-dictionary__download-link {
+					@include display(flex);
+					@include align-items(center);
+					.data-dictionary__download-icon {
+						@include display(flex);
+						@include flex(0 0 auto);
+						height: rem(15);
+						width: rem(15);
+						padding-right: rem(5);
+						.usa-da-spreadsheet {
+							height: rem(15);
+							width: rem(15);
+						}
+					}
+					font-size: $small-font-size;
+				}
+			}
+
+			.data-dictionary__table-wrapper {
+				@import './_dictionary-table';
+			}
+			@import './_search-bar';
+		}
+	}
+}
